Add google maps for the main screen. Display places with pins in the map.

+++ app/src/main/java/com/szabolcs/musicbrainz/data/model/remote/Coordinates.kt
@@ -3,6 +3,6 @@ package com.szabolcs.musicbrainz.data.model.remote
 import com.google.gson.annotations.SerializedName
 
 data class Coordinates(
-    @SerializedName("latitude") val latitude: Float,
-    @SerializedName("longitude") val longitude: Float
+    @SerializedName("latitude") val latitude: Double,
+    @SerializedName("longitude") val longitude: Double
 )
